# experiments/modelless-mcp/specs/openapi.yml
openapi: 3.0.3
info:
  title: MCP Catalog — Control Plane API (modelless)
  version: 0.1.0
  description: >
    REST control-plane API for the QA MCP **catalog** (modelless variant). This
    complements the MCP protocol surface (JSON-RPC over Streamable HTTP). It
    exposes catalog browsing, tenant configuration, asset manifests, and
    **install descriptors** for the `hard install` flow. The catalog
    **distributes** assets and **executes nothing**: it hosts no model, performs
    no inference, and runs no skills. All reasoning and skill execution happen in
    the CLIENT's local model. Authentication is Azure AD / OIDC (OAuth2). All
    endpoints except /health require a valid Bearer token.
  contact:
    name: QA Platform Team
  license:
    name: Proprietary
servers:
  - url: https://mcp.example-corp.internal/v1
    description: Corporate internal endpoint (behind API Management)
security:
  - azureAd: []
...
paths:
...
  /tenants/{tenantId}/overrides:
    get:
      tags: [tenants]
      summary: Get tenant overrides (customization)
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Overrides
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantOverrides' }
    put:
      tags: [tenants]
      summary: Upsert tenant overrides (admin role required)
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TenantOverrides' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantOverrides' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    azureAd:
      type: oauth2
      description: Azure AD (Entra ID) OAuth2 / OIDC
      flows:
        clientCredentials:
          tokenUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/token
          scopes:
            api://mcp-server/.default: Default access scope
  parameters:
...
    TenantId:
      in: path
      name: tenantId
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    Forbidden:
      description: Authenticated but not authorized (RBAC)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    TenantOverrides:
      type: object
      description: >
        Tenant customization. There is NO preferred_model and NO token quota,
        because the server performs no inference.
      properties:
        tenant_id: { type: string, format: uuid }
        style: { type: object, additionalProperties: true }
        skill_params:
          type: object
          additionalProperties: true
        quotas:
          type: object
          properties:
            requests_per_min: { type: integer }
    Error:
      type: object
      properties:
        code: { type: string, example: forbidden }
        message: { type: string }
        request_id: { type: string }